+++ meta_package_manager/pool.py
@@ -98,7 +98,7 @@ class ManagerPool:
     """A dict-like register, instanciating all supported package managers."""
 
     ALLOWED_EXTRA_OPTION: Final = frozenset(
-        {"ignore_auto_updates", "stop_on_error", "dry_run"},
+        {"ignore_auto_updates", "stop_on_error", "dry_run", "timeout"},
     )
     """List of extra options that are allowed to be set on managers during the use of
     the :py:func:`meta_package_manager.pool.ManagerPool.select_managers` helper
@@ -185,7 +185,7 @@ class ManagerPool:
         keep_unsupported: bool = False,
         drop_inactive: bool = True,
         implements_operation: Operations | None = None,
-        **extra_options: bool,
+        **extra_options: bool | int,
     ) -> Iterator[PackageManager]:
         """Utility method to extract a subset of the manager pool based on selection
         list (``keep`` parameter) and exclusion list (``drop`` parameter) criterion.